Create AllowCredentials.bb
This profile creates a passive issue for the string "Access-Control-Allow-Credentials: true".  It is set as informational.  

While Burp Scanner will attempt to manipulate the origin header, it may miss vulnerabilities where you cannot supply any origin and have it be reflected, but can manipulate the origin in specific ways to trigger a CORS bypass.